fix questions, fetch questionniare status on login
