+++ src/components/questionnaire/questions/definitions/VarigFritid1QuestionDefinition.ts
@@ -5,7 +5,7 @@ export const VARIGFRITID1: QuestionFormType = {
     id: 'varigfritid1',
     order: 117,
     theme: Theme.rarePurchases,
-    questionText: 'Fritidsutstyr, Har du/dere kjøpt i løpet av de siste 12 månedene kjøpt eller fått… ',
+    questionText: 'Har du/dere i løpet av de siste 12 månedene kjøpt eller fått… ',
     hasAnswered: false,
     answerValue: {
         answers: [
